Refactor: move to own module
