+++ services/galley/galley.cabal
@@ -31,12 +31,14 @@ library
     Galley.API.Mapping
     Galley.API.Message
     Galley.API.MLS
+    Galley.API.MLS.GroupInfo
     Galley.API.MLS.KeyPackage
     Galley.API.MLS.Keys
     Galley.API.MLS.Message
     Galley.API.MLS.Propagate
     Galley.API.MLS.Removal
     Galley.API.MLS.Types
+    Galley.API.MLS.Util
     Galley.API.MLS.Welcome
     Galley.API.One2One
     Galley.API.Public
